test: Refactor decompilation script testing framework and use lit runner.
