--- scripts/ghidra/test/src/argc.c
@@ -1,8 +1,0 @@
-// CHECK: {{...}}
-int argc(int argc, char **argv) {
-    return argc;
-}
-int main(int a, char **argv)
-{
-    return argc(a, argv);
-}
