--- scripts/ghidra/test/src/argc.function
@@ -1,1 +1,0 @@
-main
