--- scripts/ghidra/test/src/array.function
@@ -1,1 +1,0 @@
-main
