--- scripts/ghidra/test/src/concat.function
@@ -1,1 +1,0 @@
-string_concat
