--- scripts/ghidra/test/src/fib.function
@@ -1,1 +1,0 @@
-fibonacci
