--- scripts/ghidra/test/src/fread.function
@@ -1,1 +1,0 @@
-main
