--- scripts/ghidra/test/src/fwrite.function
@@ -1,1 +1,0 @@
-write_file
