--- scripts/ghidra/test/src/insert.function
@@ -1,1 +1,0 @@
-insert_substring
