--- scripts/ghidra/test/src/list.function
@@ -1,1 +1,0 @@
-print_list
