--- scripts/ghidra/test/src/matrix.function
@@ -1,1 +1,0 @@
-multiply_matrices
