--- scripts/ghidra/test/src/prime.function
@@ -1,1 +1,0 @@
-is_prime
