--- scripts/ghidra/test/src/queue.function
@@ -1,1 +1,0 @@
-dequeue
