--- scripts/ghidra/test/src/reverse.function
@@ -1,1 +1,0 @@
-reverse_string
