--- scripts/ghidra/test/src/sort.function
@@ -1,1 +1,0 @@
-main
