--- scripts/ghidra/test/src/struct.function
@@ -1,1 +1,0 @@
-main
