--- scripts/ghidra/test/src/structb.function
@@ -1,1 +1,0 @@
-main
