--- scripts/ghidra/test/src/sub.c
@@ -1,7 +1,0 @@
-// CHECK:  {{...}}
-int main()
-{
-	int x;	
-	x = 4;
-	return x - 4;
-}
