--- scripts/ghidra/test/src/sub.function
@@ -1,1 +1,0 @@
-main
