--- scripts/ghidra/test/src/union.function
@@ -1,1 +1,0 @@
-main
