--- test/ghidra.cpp
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2024, Trail of Bits, Inc.
- * All rights reserved.
- *
- * This source code is licensed in accordance with the terms specified in
- * the LICENSE file found in the root directory of this source tree.
- */
-auto main() -> int { return 0; }
