+++ test/ghidra/fread.c
@@ -1,6 +1,12 @@
+// UNSUPPORTED: system-windows
+// RUN: %cc %s -o %t
+// RUN %t; if [ "$(uname)" = "Linux" ]; then %decompile-headless %t main %t1 fi
+// RUN %t; if [ "$(uname)" = "Darwin" ]; then %decompile-headless %t _main %t1 fi
+// RUN %t1; %file-check %s --input-file %t1
+// CHECK: {{...}}
+
 #include <stdio.h>
 
-//CHECK: {{...}}
 int main(void) {
     FILE *file = fopen("example.txt", "r");
     if (!file) {
