+++ test/ghidra/fwrite.c
@@ -1,6 +1,12 @@
+// UNSUPPORTED: system-windows
+// RUN: %cc %s -o %t
+// RUN %t; if [ "$(uname)" = "Linux" ]; then %decompile-headless %t write_file %t1 fi
+// RUN %t; if [ "$(uname)" = "Darwin" ]; then %decompile-headless %t _write_file %t1 fi
+// RUN %t1; %file-check %s --input-file %t1
+// CHECK: {{...}}
+
 #include <stdio.h>
 
-// CHECK: {{...}}
 void write_file(const char* filename, const char* content) {
     FILE* file = fopen(filename, "w");
     if (file == NULL) {
