+++ test/ghidra/reverse.c
@@ -1,7 +1,13 @@
+// UNSUPPORTED: system-windows
+// RUN: %cc %s -o %t
+// RUN %t; if [ "$(uname)" = "Linux" ]; then %decompile-headless %t reverse_string %t1 fi
+// RUN %t; if [ "$(uname)" = "Darwin" ]; then %decompile-headless %t _reverse_string %t1 fi
+// RUN %t1; %file-check %s --input-file %t1
+// CHECK: {{...}}
+
 #include <stdio.h>
 #include <string.h>
 
-// CHECK: {{...}}
 void reverse_string(char *str) {
     int len = strlen(str);
     for (int i = 0; i < len / 2; ++i) {
