+++ test/ghidra/structb.c
@@ -1,0 +1,26 @@
+// UNSUPPORTED: system-windows
+// RUN: %cc %s -o %t
+// RUN %t; if [ "$(uname)" = "Linux" ]; then %decompile-headless %t main %t1 fi
+// RUN %t; if [ "$(uname)" = "Darwin" ]; then %decompile-headless %t _main %t1 fi
+// RUN %t1; %file-check %s --input-file %t1
+// CHECK: {{...}}
+
+#include <stdio.h>
+
+typedef struct {
+    char name[50];
+    int age;
+} Person;
+
+// CHECK: {{...}}
+int main(void) {
+    Person p;
+    snprintf(p.name, sizeof(p.name), "John");
+    p.age = 30;
+
+    printf("Name: %s\n", p.name);
+    printf("Age: %d\n", p.age);
+
+    return 0;
+}
+
